Make CalciteSchema extendable for different implemetations by refactoring CalciteSchema and making it an interface
