+++ core/src/main/java/org/apache/calcite/jdbc/SimpleCalciteRootSchema.java
@@ -14,13 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.calcite.jdbc;
 
+import org.apache.calcite.schema.Schema;
 
 /**
- * Interface that marks the root schema.
+ * The base level schema (no parent) for SimpleCalciteSchema
  */
-public interface CalciteRootSchema extends CalciteSchema {
-}
 
-// End CalciteRootSchema.java
+public class SimpleCalciteRootSchema extends SimpleCalciteSchema
+    implements CalciteRootSchema {
+
+  /** Creates a root schema. */
+  SimpleCalciteRootSchema(Schema schema) {
+    super(null, schema, "");
+  }
+
+}
