+++ core/src/main/java/org/apache/calcite/prepare/PlannerImpl.java
@@ -17,7 +17,7 @@
 package org.apache.calcite.prepare;
 
 import org.apache.calcite.adapter.java.JavaTypeFactory;
-import org.apache.calcite.jdbc.CalciteSchema;
+import org.apache.calcite.jdbc.CalciteSchemaImpl;
 import org.apache.calcite.plan.RelOptCluster;
 import org.apache.calcite.plan.RelOptPlanner;
 import org.apache.calcite.plan.RelOptSchema;
@@ -224,9 +224,9 @@ public class PlannerImpl implements Planner {
                               SchemaPlus rootSchema, List<String> schemaPath) {
       // View may have different schema path than current connection.
       final CalciteCatalogReader catalogReader = new CalciteCatalogReader(
-          CalciteSchema.from(rootSchema),
+          CalciteSchemaImpl.from(rootSchema),
           parserConfig.caseSensitive(),
-          CalciteSchema.from(defaultSchema).path(null),
+          CalciteSchemaImpl.from(defaultSchema).path(null),
           typeFactory).withSchemaPath(schemaPath);
       return expandViewHelper(queryString, catalogReader);
     }
@@ -271,9 +271,9 @@ public class PlannerImpl implements Planner {
   private CalciteCatalogReader createCatalogReader() {
     SchemaPlus rootSchema = rootSchema(defaultSchema);
     return new CalciteCatalogReader(
-        CalciteSchema.from(rootSchema),
+        CalciteSchemaImpl.from(rootSchema),
         parserConfig.caseSensitive(),
-        CalciteSchema.from(defaultSchema).path(null),
+        CalciteSchemaImpl.from(defaultSchema).path(null),
         typeFactory);
   }
